Refactor the aggregator
Put corresponding logic to the right places.
